refactor : ChessGame 패키지 domain으로 변경
