+++ src/main/java/chess/App.java
@@ -1,6 +1,7 @@
 package chess;
 
 import chess.controller.ChessGameController;
+import chess.domain.ChessGame;
 
 public class App {
     public static void main(String[] args) {
